refactors filters, adds fetch and store exchange rates functionalities and tests, updates README file, adds api documentation

+++ database/migrations/2025_07_18_090333_exchange_rate.php
@@ -11,7 +11,7 @@ return new class extends Migration {
             $table->id();
             $table->string('currency_from')->default('EUR');
             $table->string('currency_to')->index();
-            $table->decimal('rate', 16, 8);
+            $table->float('rate');
             $table->dateTime('retrieved_at');
             $table->timestamps();
         });
